+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to the extension.
 
 ### Added
 
+- Now auto-completes object properties from expressions like `var['prop'] = { prop1: 2 }`.
 - SugarCube: Macros inside link setters (such as `[[passage][<<set $var to 1>>]]`) now generate a warning about them not being evaluated.
 
 ### Changed